Revert "move relay hub penalization util functions to test utils file"
This reverts commit dd36b147256a90ac5a7bb9983188762d36a451bd.
